feat: use base evaluator
