+++ dynamiq/evaluations/__init__.py
@@ -1,1 +1,2 @@
+from .base_evaluator import BaseEvaluator
 from .python_evaluator import PythonEvaluator
