+++ dynamiq/evaluations/metrics/answer_correctness.py
@@ -1,14 +1,11 @@
 import json
-import logging
-from functools import cached_property
 
-from pydantic import BaseModel, Field, PrivateAttr, computed_field, field_validator, model_validator
+from pydantic import BaseModel, Field, PrivateAttr, field_validator, model_validator
 
 from dynamiq.components.evaluators.llm_evaluator import LLMEvaluator
+from dynamiq.evaluations import BaseEvaluator
 from dynamiq.nodes.llms import BaseLLM
-
-# Configure logging
-logger = logging.getLogger(__name__)
+from dynamiq.utils.logger import logger
 
 
 class ExtractStatementsInput(BaseModel):
@@ -140,7 +137,7 @@ class RunOutput(BaseModel):
     final_scores: list[float]
 
 
-class AnswerCorrectnessEvaluator(BaseModel):
+class AnswerCorrectnessEvaluator(BaseEvaluator):
     """
     Evaluator class for computing the correctness of answers given
     questions and ground truths.
@@ -157,14 +154,6 @@ class AnswerCorrectnessEvaluator(BaseModel):
     _statement_extractor: LLMEvaluator = PrivateAttr()
     _statement_classifier: LLMEvaluator = PrivateAttr()
     _similarity_evaluator: LLMEvaluator = PrivateAttr()
-
-    class Config:
-        arbitrary_types_allowed = True  # Allow arbitrary types like LLMEvaluator and BaseLLM
-
-    @computed_field
-    @cached_property
-    def type(self) -> str:
-        return f"{self.__module__.rsplit('.', 1)[0]}.{self.__class__.__name__}"
 
     def __init__(self, **data):
         super().__init__(**data)
