+++ dynamiq/evaluations/metrics/context_precision.py
@@ -1,13 +1,9 @@
-import logging
-from functools import cached_property
-
-from pydantic import BaseModel, PrivateAttr, computed_field, field_validator, model_validator
+from pydantic import BaseModel, PrivateAttr, field_validator, model_validator
 
 from dynamiq.components.evaluators.llm_evaluator import LLMEvaluator
+from dynamiq.evaluations import BaseEvaluator
 from dynamiq.nodes.llms import BaseLLM
-
-# Configure logging
-logger = logging.getLogger(__name__)
+from dynamiq.utils.logger import logger
 
 
 class ContextPrecisionInput(BaseModel):
@@ -78,7 +74,7 @@ class VerdictResult(BaseModel):
         return v
 
 
-class ContextPrecisionEvaluator(BaseModel):
+class ContextPrecisionEvaluator(BaseEvaluator):
     """
     Evaluator class for context precision metric.
 
@@ -90,14 +86,6 @@ class ContextPrecisionEvaluator(BaseModel):
 
     # Private attribute (not part of the Pydantic model fields)
     _context_precision_evaluator: LLMEvaluator = PrivateAttr()
-
-    class Config:
-        arbitrary_types_allowed = True  # Allow arbitrary types
-
-    @computed_field
-    @cached_property
-    def type(self) -> str:
-        return f"{self.__module__.rsplit('.', 1)[0]}.{self.__class__.__name__}"
 
     def __init__(self, **data):
         super().__init__(**data)
