+++ dynamiq/evaluations/metrics/context_recall.py
@@ -1,15 +1,12 @@
 import json
-import logging
-from functools import cached_property
 from typing import Any
 
-from pydantic import BaseModel, PrivateAttr, computed_field, field_validator, model_validator
+from pydantic import BaseModel, PrivateAttr, field_validator, model_validator
 
 from dynamiq.components.evaluators.llm_evaluator import LLMEvaluator
+from dynamiq.evaluations import BaseEvaluator
 from dynamiq.nodes.llms import BaseLLM
-
-# Configure logging
-logger = logging.getLogger(__name__)
+from dynamiq.utils.logger import logger
 
 
 class ContextRecallInput(BaseModel):
@@ -95,7 +92,7 @@ class ContextRecallOutput(BaseModel):
     final_scores: list[float]
 
 
-class ContextRecallEvaluator(BaseModel):
+class ContextRecallEvaluator(BaseEvaluator):
     """
     Evaluator class for context recall metric.
 
@@ -107,14 +104,6 @@ class ContextRecallEvaluator(BaseModel):
 
     # Private attribute
     _classification_evaluator: LLMEvaluator = PrivateAttr()
-
-    class Config:
-        arbitrary_types_allowed = True  # Allow arbitrary types
-
-    @computed_field
-    @cached_property
-    def type(self) -> str:
-        return f"{self.__module__.rsplit('.', 1)[0]}.{self.__class__.__name__}"
 
     def __init__(self, **data):
         super().__init__(**data)
