+++ dynamiq/evaluations/metrics/factual_correctness.py
@@ -1,14 +1,11 @@
-import logging
-from functools import cached_property
 from typing import Any
 
-from pydantic import BaseModel, PrivateAttr, computed_field, field_validator, model_validator
+from pydantic import BaseModel, PrivateAttr, field_validator, model_validator
 
 from dynamiq.components.evaluators.llm_evaluator import LLMEvaluator
+from dynamiq.evaluations import BaseEvaluator
 from dynamiq.nodes.llms import BaseLLM
-
-# Configure logging
-logger = logging.getLogger(__name__)
+from dynamiq.utils.logger import logger
 
 
 class DecomposeClaimsInput(BaseModel):
@@ -116,7 +113,7 @@ class RunOutput(BaseModel):
     final_scores: list[float]
 
 
-class FactualCorrectnessEvaluator(BaseModel):
+class FactualCorrectnessEvaluator(BaseEvaluator):
     """
     Evaluator class for factual correctness metric.
 
@@ -136,14 +133,6 @@ class FactualCorrectnessEvaluator(BaseModel):
 
     _claim_decomposer: LLMEvaluator = PrivateAttr()
     _nli_evaluator: LLMEvaluator = PrivateAttr()
-
-    class Config:
-        arbitrary_types_allowed = True  # Allow arbitrary types
-
-    @computed_field
-    @cached_property
-    def type(self) -> str:
-        return f"{self.__module__.rsplit('.', 1)[0]}.{self.__class__.__name__}"
 
     def __init__(self, **data):
         super().__init__(**data)
